requirements.txt
pandas
numpy
sklearn
keras
tensorflow
matplotlib
win10toast
python-binance